Refactored dummy endpoint into tests endpoint
