+++ src/env_data/tests.py
@@ -7,48 +7,4 @@
 # Contact: plantalytics.capstone@gmail.com
 #
 
-from django.test import TestCase, Client
-from django.test.utils import setup_test_environment
-
-
-class MainTests(TestCase):
-    """
-    Executes all of the unit tests for the 'env_data' endpoint.
-    """
-    def test_response_temperature_data(self):
-        setup_test_environment()
-        client = Client()
-        response = client.get('/env_data/'
-                              + '?vineyard_id=0&'
-                              + 'env_variable=temperature')
-        self.assertEqual(response.status_code, 200)
-
-    def test_response_humidity_data(self):
-        setup_test_environment()
-        client = Client()
-        response = client.get('/env_data/'
-                              + '?vineyard_id=0&'
-                              + 'env_variable=humidity')
-        self.assertEqual(response.status_code, 200)
-
-    def test_response_leafwetness_data(self):
-        setup_test_environment()
-        client = Client()
-        response = client.get('/env_data/'
-                              + '?vineyard_id=0&'
-                              + 'env_variable=leafwetness')
-        self.assertEqual(response.status_code, 200)
-
-    def test_admin_redirect(self):
-        setup_test_environment()
-        client = Client()
-        response = client.get('/admin/')
-        expected_redirect_url = '/admin/login/?next=/admin/'
-        self.assertEqual(response.status_code, 302)
-        self.assertEqual(response.url, expected_redirect_url)
-
-    def test_admin(self):
-        setup_test_environment()
-        client = Client()
-        response = client.get('/admin/login/?next=/admin/')
-        self.assertEqual(response.status_code, 200)
+# Create tests here
