+++ src/login/tests.py
@@ -6,54 +6,5 @@
 #     Please see the file LICENSE in this distribution for license terms.
 # Contact: plantalytics.capstone@gmail.com
 #
-import os
 
-from django.test import TestCase, Client
-from django.test.utils import setup_test_environment
-
-
-class MainTests(TestCase):
-    """
-    Executes all of the unit tests for the 'login' endpoint.
-    """
-    def test_response_valid_login(self):
-        setup_test_environment()
-        client = Client()
-        response = client.get('/login/'
-                              + '?username='
-                              + os.environ.get('LOGIN_USERNAME')
-                              + '&password='
-                              + os.environ.get('LOGIN_PASSWORD'))
-        self.assertEqual(response.status_code, 200)
-
-    def test_response_invalid_username(self):
-        setup_test_environment()
-        client = Client()
-        response = client.get('/login/'
-                              + '?username=mrawesome'
-                              + '&password='
-                              + os.environ.get('LOGIN_PASSWORD'))
-        self.assertEqual(response.status_code, 403)
-
-    def test_response_invalid_password(self):
-        setup_test_environment()
-        client = Client()
-        response = client.get('/login/'
-                              + '?username='
-                              + os.environ.get('LOGIN_USERNAME')
-                              + '&password=notcorrect')
-        self.assertEqual(response.status_code, 403)
-
-    def test_admin_redirect(self):
-        setup_test_environment()
-        client = Client()
-        response = client.get('/admin/')
-        expected_redirect_url = '/admin/login/?next=/admin/'
-        self.assertEqual(response.status_code, 302)
-        self.assertEqual(response.url, expected_redirect_url)
-
-    def test_admin(self):
-        setup_test_environment()
-        client = Client()
-        response = client.get('/admin/login/?next=/admin/')
-        self.assertEqual(response.status_code, 200)
+# Create tests here
