+++ src/tests/apps.py
@@ -10,5 +10,5 @@
 from django.apps import AppConfig
 
 
-class DummyConfig(AppConfig):
-    name = 'dummy'
+class TestsConfig(AppConfig):
+    name = 'tests'
